+++ app/Http/Controllers/ProfileController.php
@@ -3,8 +3,8 @@
 namespace App\Http\Controllers;
 
 use App\Enums\ResponseCodeEnum;
+use App\Service\Profile\ProfileService;
 use Illuminate\Http\Request;
-use App\Models\ProfileModel;
 use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\ProfileAvatarRequest;
 use Validator;
@@ -17,6 +17,19 @@ use Illuminate\Http\Response;
 class ProfileController extends Controller
 {
 
+
+    /** @var ProfileService */
+    protected $profileService;
+
+    /**
+     * ProfileController constructor.
+     * @param ProfileService $profileService
+     */
+    function __construct(ProfileService $profileService)
+    {
+        $this->profileService = $profileService;
+    }
+
     /**
      * @param int $id
      * @return \Illuminate\Contracts\Foundation\Application|
@@ -25,10 +38,10 @@ class ProfileController extends Controller
      */
     public function viewAnotherProfile(int $id)
     {
-        $userData = ProfileModel::getAnotherUser($id);
 
-        if (!empty($userData->name)) {
-            return view('menu.profile', ['data_user' => $userData]);
+        $anotherUserData = $this->profileService->getAnotherUser($id);
+        if (!empty($anotherUserData->name)) {
+            return view('menu.profile', ['data_user' => $anotherUserData]);
         } else {
             return view('error_404', ['error' => ['Данного пользователя не существует']]);
         }
@@ -43,7 +56,7 @@ class ProfileController extends Controller
     public function viewProfile(Request $request)
     {
 
-        $userData = ProfileModel::getUserData();
+        $userData = $this->profileService->getUserData();
         return view('menu.profile', ['data_user' => $userData]);
     }
 
@@ -56,7 +69,7 @@ class ProfileController extends Controller
     public function changeProfile(int $userId)
     {
 
-        $userData = ProfileModel::getUserDataChange($userId);
+        $userData = $this->profileService->getUserDataChange($userId);
         return view('menu.profile.change_profile', ['data_user' => $userData]);
     }
 
@@ -69,7 +82,7 @@ class ProfileController extends Controller
 
         if ($request->ajax()) {
             $input = $request->only(['data_send']);
-            $back = ProfileModel::changeProfile($input);
+            $back = $this->profileService->changeProfile($input);
 
             try
             {
@@ -93,8 +106,8 @@ class ProfileController extends Controller
      */
     public function changeAvatar(ProfileAvatarRequest $request)
     {
-        $isChanged = ProfileModel::changeAvatar($request);
 
+        $isChanged = $this->profileService->changeAvatar($request);
         return redirect()->route('profile_id', Auth::user()->id)
             ->with('success', 'Вы успешно изменили аватар')
             ->with('isChanged', $isChanged);
